BUG: Fixing Regression introduced by !2514
The issue is that GroupItem::hasRelevantChildren was always checking against the set of active categories,
but in the case of an Analysis Attribute, we ignore categories all together.  The fix was to provide an option to skip the category check in the method.

This also required isRelevant to have the same change.

Also noticed that isIgnore was not being taken into consideration via the isValid method.  Since this test should be done when isEnbaled is being checked I noticed that the isEnabled check was being done in isValidInternal
(which meant each Item class was copying the same code fragment) and that it was being done after the category check
which is more expensive.

This has been refactored to the isValid is doing the isIgnored and isEnabled tests.

Also update unitIsRelevant test to include a test for hasRelevantChildren.