BUG: qtAttributeView reporting incorrect validity
The original logic used the visibility state of the association widget to determine
if it's validity should factor into the qtAttributeView that owned the widget. Though this
does work when the association widget is being initially constructed, it fails when the
widget is being updated though its operation handler. My guess is that in this case the widget is
currently hidden and hence its visibility method was returning false.

The fix was to store a member variable to indicate if the association widget is currently in use.