ENH: Adding a fast way to find DescriptivePhrases of Persistent Objects
Added a map between a Persistent Object's id and a set of DescriptivePhases that contains it.
This should speed up updating the PhraseModel when Persistent Objects get created or expunged.

Also changed the signature of SubphraseGenerator::indexOfObjectInParent to take in a const smtk::view::DescriptivePhrasePtr& instead of a smtk::view::DescriptivePhrasePtr& for the parent Sub-phrase.