gitlab-ci: move job name prefixes to be suffixes
Due to the way GitLab renders job names, common prefixes end up
obscuring the "important" parts of names by pushing them off of the
bubbles and requiring a tooltip to distinguish between similar jobs.
With this naming, jobs with similar prefixes are more closely related to
each other and any rendering of GitLab will either have them in
different stages or deeper in the dependency graph.