Performance: Reverting SubphraseGenerator::subphrasesForCreatedObjects
It seems that the calculation of a Model Entity's Bordants (which is needed
to calculate the parents of the entity) is more costly that traversing the Phase Model
tree to determine where to insert the new phrases.

Doing this as a commit instead of an amend so we have a record of the new algorithm which
could be used if the Bordant calculation gets improved in the future.