Handle GroupItem in attribute_builder and python operation tracing.
Add handling in `Attribute::itemAtPath()` for group indices - a path
can contain a `#N` element between a group item and its children to
specify which sub-group they belong to.

Modify python operation tracing and AttributeBuilder to generate and
handle the new syntax. Include a "count" keyword for groups so we
don't have to pre-parse paths to guess the number of groups.