Group item paths use an index, add python tracing test
Group item paths can include an integer for their sub-group index.
If an item belonging to a group has a name that is an integer, the
sub-group index must be used to retrieve that item using `itemAtPath()`

Move and wrap pqSMTKPythonTrace so it can be used in a python test,
and test whether it correctly traces group items in a test operation.