pybind: fix hashable requirements
The `CellEntity` type is hashable through its `EntityRef` base class, so
add `__hash__` there.

For `MeshSet`, however, the cell and point field methods return
`std::set` which only requires `operator<`. The Python `set` requires
its values be hashable. This is not the case for the `CellField` or
`PointField` types. Instead, return vectors of the values.