Improve python tracing of Import operations
Add event processing to pqSMTKBehavior, and call in the trace, so that
the pipeline source is available when ParaView adds a call to
GetActiveSource(). This pairs with an app change that blocks tracing
the standard ParaView reader in favor of an SMTK operation.

Fix a bug in replay that crashed when the resource list was empty.

Fix a bug tracing operations with extensible items with one value.