Add a setting to allow 3D widgets to show when their Qt gui hides.
The default behavior of ParaView 3D widgets is to hide when their
Qt control panel looses focus or is hidden. Add a toggle
`pqSMTKAttributeItemWidget::setHideWidgetWhenInactive()` that allows
an app to choose to always show 3D widgets, and have the user
toggle them on an off explicitly. Setting is propogated to
`pqSMTKInteractivePropertyWidget` base class for property widgets,
which overrides ParaView hiding behavior when needed.