Add a template type and version number to attribute::Resource.
The type and version number are for use by SBT developers who need
to version the attribute and item definitions for use in migrating
instanced attributes to a newer SBT schema than was used to create
a resource in the first place.

Note that the changes to the PEGTL parsing code are required
once the attribute resource includes `smtk/string/Token.h` because
that defines a namespace (`string`) which makes the PEGTL
`string<>` template ambiguous. Use the macro PEGTL provides instead.