Improvements for operation views.
Sometimes custom items need access to the operation view which created
them. This was not possible with qtOperationView because internally it
created a qtInstancedView which owned the item. So, we add the operation
view to the instanced-view's config-information type-container and add
a method so items can fetch the config-information type-container from
their parent view.

The qtOntologyItem is an example of a qtItem intended specifically for
operation views that needs to reset its state when its operation
completes (and it needs to own the lock on the operation parameters in
order to do this so it does not misconfigure an operation while
running).

Provide a `disableApplyAfterRun` property for the view which
custom item views may set to indicate that re-running an operation
with identical parameters is valid (or, in the case of
qtOntologyItem, indicates the item cannot know when users have
altered parameters).