ENH: Added ability for Instanced Views to find Attributes by ID
You can now specify the ID of the Attribute by adding a View Component Attributed called ID whose
value is the Attribute's UUID (as a string).

The previous methods of finding the Attribute by Name or using Name and Type to create the Attribute will still work;
however, the View will modify its configuration by adding the ID attribute set to the Attribute that was found by Name (or created using Name and Type).  This way the View would still be appropriate even if the Attributes' Names where changed afterwords.