BUG: Fixing Instance View Issue
When the ID support was added, the including making a smtk::view::Configuration::Component variable a component so that it could be modified by adding the newly created Attribute's ID.  Unfortunately the same variable is later assigned to a style which inadvertently changed the configuration making it invalid.

This change undoes the reference change and directly modified the configuration with the ID information.