Do not automatically manage projects as they are created.
Instead, have the project manager observe operations which
create projects and manage any new projects upon completion;
this commit changes when projects are added to the `project::Manager`.
This matches the pattern set by the resource manager and
avoids observers being fired during operations when the
project may not be in a valid state.

Before, this was done at creation; now it is done at the end
of observing operation results. This ensures the project is
not empty when project-manager observers are fired (unless the
project is "blank").

This commit also
+ redirects project and task-manager observers to the GUI thread so
  that the task panel UI's observers run in the proper thread.
+ does not use a QTimer to asynchronously raise the task panel but
  instead raises it immediately in the observer (now that the observer
  is run in the GUI thread).