BUG: Fix Issue with displaying values with unknown units
There was a redundant setText call when creating a widget for a value
with either no or unknown units

In addition, setting the widget for an item with unknown units would try to
use the item's "raw" string representation which included both the value and unit.
This would cause the double validator to convert the entire string to 0.

Now the string is properly split and only the value portion is used to set the
widget.