BUG: Fixing Crash due to Widget destruction
Changed the connection type from Queued to Direct when editing is
completed w/r to both unit and non-unit line edit widgets.

The issue was that a queued connection may be processed after the widget
that caused the signal to be fired is deleted.

Also corrected an issue when resetting an item's value back to its default when the item's
units are unknown to the units system.  In this case the widget must only be giving the default
value string without the units attached.