ENH: Added a readOnly mode to the UIManager
Added API to set the qtUIManager to be read-only.  Setting this to be true will
tell all the Views being displayed by the manager that no modifications should be
allowed.

This mechanism is now being used when a Task is marked as Completed and the contents of the Attribute Panel needs to be read-only.  This change will allow the user to browse multiple tabs if the Task style is using a Group Tab View.

Other changes:

1. 3D based Item Views now will honor isReadOnly "property"
2. qtDefaultTaskNode's menus now will change Mark completed to Unmark as completed when a Task as been marked as completed.
3. The Attribute Editor Panel will now reset and be blank if the current Task does not require it.

Developer changes
~~~~~~~~~~~~~~~~~~

Added the following methods:

qtUIManager::setReadOnly(bool val)
bool qtUIManager::isReadOnly() const