Simplify the qtDefaultTaskNode.
This includes copying the previous UI file that was shared by both
example task node (`qtDefaultTaskNode1`) so that the remaining task
node can continue to use it.

By changing `qtBaseTaskNode::updateTaskState()` to accept the
active/inactive state, we avoid querying the task-manager for
the active task in the middle of a transition, which was causing
the UI to be inconsistent.