Remove an invalid optimization.
When a resource is loaded with a task marked as active,
the task is active before the panel is listening for active
tasks. This causes the active-task observer to be fired
immediately, which calls the observer with the current task
in both the "old" and "new" active task parameters. So we
cannot skip this case.