BUG: Fix Crash when modifying an Analysis Attribute
While using the ConfigurationTest.sbt file, a crash was found when modifying the attribute.  The issue is that qtAttribute emits 2 signals (one indicating the item that was modified and one that indicates that the attribute was modified).  In this case, the first signal indirectly caused the qtAttribute to be deleted so that the second signal caused a crash).  This change checks to see if the qtAttribute has been deleted before sending the second signal.

Looking into possibly removing the second signal all together in the long term.