ENH: Allowing Analysis Attributes to be Defined in a Template File
This change allows a template file to include a Configuration section that
defines the Analysis Attribute Definition as well as an Attributes section that
instantiates Attributes based on the Analysis Attribute Definition.

Previously the Configuration Section was processed after the Attribute Section in the XML
which made this not possible.

The analysisConfigrationExample.sbt file has been modified to show this is now possible.