Fix windows type-name issue.
Microsoft adds non-sensical cruft to the type-names it reports
unlike every other platform, even including template parameters.
Remove it thoroughly, not just when it is at the front of the
type-name.

When anonymous namespaces are involved, replace the fonky tick
marks in "`anonymous namespace'" with open/close-parentheses
like every other platform.