Add mode buttons to the task panel.
We add "pan" mode, "select" mode, and "connect" mode.

In "connect" mode, users may select a type of connection
to create and then click nodes to connect them with an
arc of the selected type. An instance of the new
qtPreviewArc class is used to show users the arc to be created
and the Escape key will abandon arc creation.

Arc types are registered by providing an arc type-name to
an operation in the `ArcCreator` operation group (a subclass
of `smtk::operation::Group`).

When nodes are moved in the the task panel, the resource
owning the task (a project) is marked as dirty (in need of
a "Save") by launching a MarkModified operation (a new operation).