ReadResource should invoke its readers internally.
This changes how the ReadResource operation invokes the sub-reader
operations it creates. Instead of invoking them synchronously as
full-fledged operations (which runs preprocessing and postprocessing
on them, including the firing of observers), it should run them
internally (passing an instance the protected Key struct) such that
no observers are fired. This way, newly-created resources are not
reported to operation observers twice (once by the sub-reader and
once by ReadResource).