Refactor deletion-via-operation-group.
Code that was in qtResourceBrowser is now a free function
to apply an operation-group to a set of objects; this
paves the way for the task-editor/resource-diagram and
others to re-use the code.

This also fixes an issue with the original approach in
qtResourceBrowser where deletion could not occur unless all
the selected objects could use the same deletion operation.