Fix the diagram's adjustment for fitting things into view.
Previously, the qtDiagram would zoom a small amount after each
operation completed; this was caused by improperly adjusting the
size of the widget.

Instead of using a fixed padding (and an incorrect one),
scale the adjustments to "fitInView()" by twice the frame
width (which appears to be the cause of the issue).