ENH: Adding Custom Filter Support for Attributes
Properties on Definitions can now be *inherited* by Attributes and derived Definitions when creating queries and association rules.

For example if Definition **A** has a floating-point property "alpha" with value 5.0, and if Definition **B** is derived from **A** and Attribute **a** is from **A** and Attribute **b** is from **B**, then all would match the rule "any[ floating-point { 'alpha' = 5.0 }]".  If later **B** was to also have floating-point property "alpha" with value 10.0 associated with it, it would override the value coming from **A** so both it and **b** would no longer pass the rule.  Similarly, properties on Attributes can override the values coming from its Definition.

Also implemented differentiation between attributes and definitions.  If the rule starts with *attribute* then only attributes have the possibility of matching the rest of the rule. Similarly, if the rule starts with *definition* then only definitions have the possibility of matching the rest of the rule.  If the rule starts with *any* or * then either attributes or definitions are allowed.

**Note** that the ``properties()`` methods on attribute and definition objects do not return *inherited* values but instead only those values local to the object. In the future, we may add methods to interact directly with inherited properties but for now only filter-strings process inherited properties.

See smtk/attribute/testing/c++/unitPropertiesFilter.cxx and data/attribute/propertiesFilterExample.sbt for examples.

Also added additional debugging methods to help track down clone resource test intermittent failures.